+++ spring-api-services/src/main/java/io/riggo/data/repositories/LoadStopSummaryRepository.java
@@ -12,5 +12,5 @@ import java.util.Optional;
 public interface LoadStopSummaryRepository extends CrudRepository<LoadStopSummary, String> {
 
     @Query(nativeQuery = true, value = "select CONCAT(ls.type,'-',ls.arrival_status) as id, ls.type, ls.arrival_status, count(*) as count from load_stop ls where ls.arrival_status is not null group by ls.type, ls.arrival_status")
-    Optional<List<LoadStopSummary>> findStopSummaryBySiteIdShipperId(Long siteId, Long shipperId);
+    Optional<List<LoadStopSummary>> findStopSummaryBySiteIdShipperId(Integer siteId, Integer shipperId);
 }
